// src/API/query/getProductConfig/__generated__/getProductConfig.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getProductConfig
// ====================================================

export interface getProductConfig_getProductConfig {
  __typename: "ProductConfig";
  productId: string;
  credit: number;
}

export interface getProductConfig {
  getProductConfig: getProductConfig_getProductConfig[];
}